# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NURTW Membership and Vehicle Verification System",
    "version": "0.1.0",
    "description": "Internal administration API for the National Union of Road Transport Workers, Anambra State Council.\n\nA verification enquiry answered by this system confirms only that a matching NURTW record exists under the requested criteria. It is not evidence of ownership, roadworthiness, licensing, or insurance, and no response should be presented as such.\n\nThis document is generated from the running application: the route list is read from the dependency-injection container and request bodies from the same schemas the API validates against. It cannot describe a route that does not exist."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nurtw_session",
        "description": "Opaque session token issued by POST /auth/login and set as an httpOnly cookie. Revocation takes effect on the next request."
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/master-data/{collection}": {
      "get": {
        "summary": "List a master-data collection.",
        "description": "Collections are `vehicle-categories` and `designations`. Designations are deliberately empty until the Union supplies them: the legacy export carries no designation list, and inventing one would present values with the appearance of Union authority (PRD §23.4).\n\n**Permission required:** `master_data.read`.",
        "operationId": "MasterDataController_list",
        "tags": [
          "MasterData"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "description": "Set to \"true\" to include deactivated entries.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Unknown collection.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie was presented.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The session does not hold master_data.read in the relevant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add an entry to a master-data collection.",
        "description": "The code is set once here and is never editable afterwards. A code is a foreign key in all but name — it appears in the legacy import mapping and in operational queries — so changing it later would be a silent data migration.\n\n**Permission required:** `master_data.manage`.",
        "operationId": "MasterDataController_create",
        "tags": [
          "MasterData"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 64,
                    "pattern": "^[A-Z][A-Z0-9_]*$"
                  },
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "sortOrder": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9999
                  }
                },
                "required": [
                  "code",
                  "label"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created."
          },
          "400": {
            "description": "Unknown collection, or the body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie was presented.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The session does not hold master_data.manage in the relevant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "That code is already in use.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/master-data/lgas/{id}": {
      "patch": {
        "summary": "Amend or deactivate a local government area.",
        "description": "The code is immutable and is not accepted here. Entries are deactivated, never deleted: member records reference them, and removing a row would orphan or block those records.\n\n**Permission required:** `master_data.manage`.",
        "operationId": "MasterDataController_updateLga",
        "tags": [
          "MasterData"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "stateName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "401": {
            "description": "No valid session cookie was presented.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The session does not hold master_data.manage in the relevant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such entry.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}